Extract validation logic into separate modules
- Created types.ts with all shared TypeScript interfaces (217 lines)
- Created validation-utils.ts with identifier validation helpers (94 lines)
- Created validation.ts with conflict detection logic (189 lines)
- Reduced refactor-engine.ts from 2477 to 2054 lines (~423 line reduction)
- Updated exports in index.ts to use types from types.ts
- All 135 tests still passing
