+++ src/refactor/engine/index.ts
@@ -1,17 +1,38 @@
 export { RefactorEngine, createRefactorEngine } from "./refactor-engine.js";
 export { WorkspaceEdit } from "./workspace-edit.js";
 export type {
-    ParserBridge,
-    SemanticAnalyzer,
-    WorkspaceReadFile,
-    WorkspaceWriteFile,
-    HotReloadUpdate,
-    ExecuteRenameRequest,
+    ApplyWorkspaceEditOptions,
+    AstNode,
+    CascadeEntry,
+    ConflictEntry,
+    DependentSymbol,
     ExecuteBatchRenameRequest,
-    RenameRequest,
-    TranspilerPatch,
+    ExecuteRenameRequest,
+    ExecuteRenameResult,
+    FileSymbol,
+    HotReloadCascadeMetadata,
+    HotReloadCascadeResult,
+    HotReloadSafetySummary,
+    HotReloadUpdate,
+    HotReloadValidationOptions,
+    MaybePromise,
+    ParserBridge,
+    PrepareRenamePlanOptions,
+    Range,
+    RefactorEngineDependencies,
     RenameImpactAnalysis,
-    ValidationSummary,
+    RenameImpactSummary,
+    RenamePlanSummary,
+    RenameRequest,
+    SemanticAnalyzer,
+    SemanticValidationResult,
+    SymbolLocation,
+    SymbolLookupResult,
+    SymbolOccurrence,
+    TranspilerBridge,
+    TranspilerPatch,
     ValidateRenameRequestOptions,
-    HotReloadSafetySummary
-} from "./refactor-engine.js";
+    ValidationSummary,
+    WorkspaceReadFile,
+    WorkspaceWriteFile
+} from "./types.js";
